requirements.txt
pydantic>=2.0
python-dateutil>=2.8
rich>=13.0
